web/requirements.txt
click==6.7
Flask==0.12.2
Flask-Cache==0.13.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
python-dateutil==2.6.1
python-rapidjson==0.0.11
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
colorama==0.3.9
colormap==1.0.1
easydev==0.9.34
cachetools==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
grpcio==1.4.0
httplib2==0.10.3
idna==2.5
monotonic==1.3
oauth2client==3.0.0
ply==3.8
protobuf==3.3.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pytz==2017.2
requests==2.18.3
rsa==3.4.2
tenacity==4.4.0
cycler==0.10.0
decorator==4.1.2
matplotlib==2.0.2
networkx==1.11
olefile==0.44
Pillow==4.2.1
pyparsing==2.2.0
PyWavelets==0.5.2

